+++ sei-ibc-go/modules/apps/27-interchain-accounts/host/types/keys.go
@@ -14,6 +14,11 @@ const (
 
 // ContainsMsgType returns true if the sdk.Msg TypeURL is present in allowMsgs, otherwise false
 func ContainsMsgType(allowMsgs []string, msg sdk.Msg) bool {
+	// check that wildcard * option for allowing all message types is the only string in the array, if so, return true
+	if len(allowMsgs) == 1 && allowMsgs[0] == "*" {
+		return true
+	}
+
 	for _, v := range allowMsgs {
 		if v == sdk.MsgTypeURL(msg) {
 			return true